Recent Customizations for iOS .vimrc
Add a line that allows customization of the cursor line background
color.

Add two leader mappings that turn spell check on and off.

Correct section where custom cursors are set based on mode.

Add a section to the Vim Lightline plugin configuration that allows
you to set the middle portion of the status bar to the same color as
the theme background (commented out by default).