Include Customizations for HTML Syntax and SnipMate
After some experimentation, I created some functions and leader
mappings that helped to get SnipMate working in my current iOS
configuration which includes an installation of both VimWiki and
SnipMate. I found that the behavior of the <Tab> key implemented by
both of these plugins was colliding and I could not expand snippets.

A stackoverflow article I read earlier today helped point me in the
direction of a solution. After working with Vim configured in this way
throughout the day, I realized that the editor had slowed down
considerably. Performance was really poor, and it was taking between
5-10 seconds to complete simple operations like moving up or down a
line. I've decided to uninstall SnipMate for now but am saving this
configuration in case I need it in the future.