Add motd used on MacBookPro
motd is an abbreviation for "Message of the Day". It is the text that is
initially displayed when you open the terminal.

The motd file on Mac OS is stored at /etc/motd

It can be viewed by entering:

cat /etc/motd

It can be modified by entering:

sudo vim /etc/motd

For more information, see:
http://osxdaily.com/2007/01/30/change-the-mac-os-x-terminals-message-of-the-day/