Scripts: Add Colortrans
I was looking for a quick way to translate the hexadecimal color values
I picked out for the AnotherKolor theme to 256 color codes that will
work in a terminal. I discovered the following gist called
`colortrans.py` on Github:

https://gist.github.com/hoov/3147229

I added this to my dotfiles scripts directory and after reading it and
testing the commands discovered that:

    colortrans (with no arguments) prints the full 256 color table.
    colortrans ff5f5f prints the xterm approximate color value.

SAMPLE OUTPUT:
RGB FF615A -> xterm color approx 203 (ff5f5f)

It even provides a nice preview of the color value and hex value.