Vim: Move Management of Custom Theme AnotherKolor to Separate Repo
Add AnotherKolor as a submodule for iOS installation.

Delete colorscheme files from Main/colors and iOS/colors in the vim
repository.

Modify the vim-setup script to install vim-anotherkolor to the
~/.vim/bundle directory.

Remove lines from the vim-setup script that create symlinks for the
AnotherKolor theme files.

Update the install-submodules script to include vim-anotherkolor.