Vim: Customize Commentary Comments for C#
Create a syntax customization for C# files by placing a cs.vim file in
the ~/.vim/after/ftplugin directory that includes an autocmd to set the
commentstring (found on Github in the vim.commentary issues).

Update the Vim installation script so it creates symlinks for the file
on desktop machines. Copy the file to the vim/iOS directory tree for iOS
installation.