Update Vim Configuration
Add a section to the PowerShell script that installs .dotfiles in
Windows so that it adds Vim configuration to the Vim 8.1 installation
directory in Windows.

Add gvim configuration settings to the .vimrc. Note that font names in
the guifont setting need to be separated by underscores intead of spaces
in Windows environments.